Reorganize layout into src, tests, examples, and scripts
