+++ examples/input_5x5_bw.png
