+++ examples/input_5x5_color.png
