+++ examples/input_9x10_bw.png
