+++ scripts/generate_example_png.py
