+++ src/nonogram_division_service/processor.py
